fixed forgot and reset password
