+++ src/app/forgotpassword/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ForgotPasswordForm from '../../components/ForgotPasswordForm';
+import ForgotPasswordForm from '../../components/authenticate/ForgotPasswordForm';
 
 const ForgotPasswordPage = () => {
     return (
