+++ src/app/register/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegisterForm from '../../components/RegisterForm';
+import RegisterForm from '../../components/authenticate/RegisterForm';
 
 const RegisterPage = () => {
   return (
