+++ src/app/resetpassword/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ResetPasswordForm from '../../components/ResetPasswordForm';
+import ResetPasswordForm from '../../components/authenticate/ResetPasswordForm';
 
 const ResetPasswordPage = () => {
     return (
